requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
httpx-sse
pyjwt[crypto]
python-dotenv
